start creating unit tests, updated textnode.py
Starting Static Site Generator (SSG), Ch2, L2. I was confused in L1 by
what they wanted the TextNode class to look like for self.text_type. As
they put it, "The type of text this node contains, which is a member of
the TextType enum." Given how they've instantiated the class in the
test provided for us, I think it's clear I'd done it incorrectly. I had
assumed the lower-case term would be passed into the class, but instead
I think it is supposed to be the Class.MEMBER notation, i.e.
TextType.LINK as opposed to "italic". really, the only change in
textnode.py is to the __repr__ method. Everything else should stay the
same... I think...