add unit tests to test_textnode.py
Added a bunch of unit tests, but also made some small changes to
main.py and textnode.py. In main.py, I added a couple of lines to call
the __eq__ function to make sure it works. In textnode.py, I added a
'return False' line to __eq__, really just to make sure it's there.