add func to textnode to convert text to htmlnode
This is C2L6 in the boot.dev Static Site course. The wording was a
little confusing to me, as it just states to define a function, but not
where to put it. I initially put it in main.py, but that seemed wrong.
I had spaced that we can define functions in modules - typing it out
now, that makes TOTAL sense, as that's the point of modules, not just
to contain classes, but also functions. After getting everything
working, I saw in the solution files the "correct" way and modified my
files.

I also made a slight change to LeafNode, in that image tags don't need
the closing tag, or the "body" text, so I added an if statement to
handle image tags specifically (hopefully this won't bite me in the
ass later).