refactor the function to copy static to public
CH5L1 - This is one of the times Boot.Dev irks me. The assignment tells
us to write a function that does x, y, and z, and so I do. Then, I look
at their solution, and y is not in the function. Their logic makes
sense, but it's not what they told us to do.

I did use some of what they did - I created global variables in main.py
to set the paths for the static and public folders. This does make
sense - the function is generic enough that it should work for any two
folders, but in this case, we are using it for two specific folders.

This also made some of the code in the function easier, as I was
initially testing this (and thus writing for) passing in just the name
of the folders, not the paths.

Lastly, I put the function in it's own python file, copy_static.py.