finish the generate function
I overachieved for once; the assignment was to generate a bunch of
pages at once in the correct folder structure. I think the intended
solution was to call the generate function multiple times, but I'm a
bit lazy, and wanted to be able to do it all at once, and so I wrote a
recursive function to do this. Turns out, the next assignment was to do
this recursively. Granted, their solution is a little more elegant than
mine, and I may eventually implement theirs, but get this - mine works!